Removing require_once where possible, adding phpstan
Starting to prepare to work with phpstan on this library, this commit
adds phpstan and removes all require_once calls where
necessary/possible so that phpstan can autoload Zend Framework through
composer (fatal errors on duplicate include of Zend_Session_Namespace
if this is not done).  A couple of other small changes to library/tests
so that phpstan runs and tests still pass with the require_onces
removed.